requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
moviepy==1.0.1
numpy==1.17.3
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.1
proglog==0.1.9
protobuf==3.10.0
pylint==2.4.3
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.22.0
scipy==1.1.0
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.36.1
typed-ast==1.4.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
